+++ guppy_examples/general/quantum-simple-1.py
@@ -1,4 +1,3 @@
-import sys
 from typing import no_type_check
 
 from guppylang import guppy, qubit
@@ -12,7 +11,3 @@ def main() -> None:
     q = qubit()
     h(q)
     result("0", measure(q))
-
-
-if __name__ == "__main__":
-    sys.stdout.buffer.write(main.compile().to_bytes())