+++ guppy_examples/quantinuum-hardware-only/rng-quantum-rng-1.py
@@ -1,4 +1,3 @@
-import sys
 from typing import no_type_check
 
 from guppylang import guppy, qubit
@@ -20,7 +19,3 @@ def main() -> None:
     r.discard()
     result("0", measure(q0))
     result("1", measure(q1))
-
-
-if __name__ == "__main__":
-    sys.stdout.buffer.write(main.compile().to_bytes())