+++ python/tests/hugr_generation.py
@@ -1,28 +1,43 @@
 import hashlib
+import importlib.util
 import json
-import subprocess
 import sys
 from dataclasses import dataclass
 from pathlib import Path
 
 import guppylang
 import hugr
+from mypy.moduleinspect import ModuleType
+
+
+def load_py_module_from_path(module_name: str, file_path: str) -> ModuleType:
+    # 1. Create a module spec from the file location
+    spec = importlib.util.spec_from_file_location(module_name, file_path)
+    msg = "Bad file path"
+    if not spec:
+        raise (ValueError(msg))
+
+    # 2. Create a new module based on that spec
+    module = importlib.util.module_from_spec(spec)
+    # 3. Optional: Add it to sys.modules so it behaves like a normal import
+    sys.modules[module_name] = module
+    # 4. Execute the module to make its functions available
+    loader = spec.loader
+    if loader:
+        loader.exec_module(module)
+    return module
+
 
 GUPPY_EXAMPLE_HASHES = (
     Path(__file__).parent / "resources/guppy_example_hugrs/hashes.json"
 )
+THIS_FILE = Path(__file__)
 
 
 def guppy_to_hugr_binary(guppy_file: Path) -> bytes:
-    result = subprocess.run(  # noqa: S603
-        [sys.executable, guppy_file],
-        check=True,
-        capture_output=True,
-        text=False,
-    )
-    hugr = result.stdout
-    assert isinstance(hugr, bytes)
-    return hugr
+    guppy_example = load_py_module_from_path("guppy_example_mod", str(guppy_file))
+    hugr_package = guppy_example.main.compile()
+    return hugr_package.to_bytes()
 
 
 def get_file_hash(path: Path) -> str:
@@ -73,11 +88,17 @@ def generate_guppy_example_dict(guppy_files: list[Path]) -> dict[str, GuppyExamp
     example_dict: dict[str, GuppyExample] = {}
     hashes = get_file_hashes()
     versions = hashes.get("guppylang_hugr_version", "")
+    this_file_hash_old = hashes.get("hugr_generation_module", "")
     current_versions = get_guppylang_hugr_version_str()
-    if versions != current_versions:
-        # if guppylang and/or hugr versions don't match regenerate everythin
+    this_file_hash = get_file_hash(THIS_FILE)
+    if versions != current_versions or this_file_hash != this_file_hash_old:
+        # if guppylang and/or hugr versions don't match
+        # or this file changed regenerate everything
         hashes = {}
-    new_hashes = {"guppylang_hugr_version": current_versions}
+    new_hashes = {
+        "guppylang_hugr_version": current_versions,
+        "hugr_generation_module": this_file_hash,
+    }
     for guppy_file in guppy_files:
         file_stem = guppy_file.stem
         hugr_file_path = (